Work around error dialogs added
Temporarily disable showing dialogs, to prevent the errors that occur
because of stopping processes with open files in the storage.
This is done by exporting a new method from the ActivityManager service.
Alternatively, we could use its configuration system, but it would be
too much overhead and could lead to other apps/services reading
incorrect configs.